Added test for invalid signing algorithm in GCP IIT plugin and applied PR comments
