+++ pkg/server/plugin/nodeattestor/gcp/iit.go
@@ -3,15 +3,10 @@ package gcp
 import (
 	"context"
 	"crypto/x509"
-	"encoding/json"
-	"encoding/pem"
 	"fmt"
-	"io/ioutil"
-	"net/http"
 	"net/url"
 	"path"
 	"sync"
-	"time"
 
 	"github.com/hashicorp/hcl"
 
@@ -29,59 +24,6 @@ const (
 
 type tokenKeyRetriever interface {
 	retrieveKey(token *jwt.Token) (interface{}, error)
-}
-
-type googlePublicKeyRetriever struct {
-	certificates map[string]*x509.Certificate
-	expirey      int64
-	mtx          *sync.Mutex
-}
-
-func (r *googlePublicKeyRetriever) retrieveKey(token *jwt.Token) (interface{}, error) {
-	if token.Header["kid"] == nil {
-		return nil, fmt.Errorf("Missing kid in identityToken header. Cannot verify token")
-	}
-	kid := token.Header["kid"].(string)
-	if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
-		return nil, fmt.Errorf("Unexpected signing method: %v", token.Header["alg"])
-	}
-
-	if r.expirey == 0 || time.Now().Unix() > r.expirey {
-		r.mtx.Lock()
-		defer r.mtx.Unlock()
-		err := r.downloadCertificates()
-		if err != nil {
-			return nil, err
-		}
-	}
-	return r.certificates[kid].PublicKey, nil
-}
-
-func (r *googlePublicKeyRetriever) downloadCertificates() error {
-	resp, err := http.Get(googleCertURL)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	bytes, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return err
-	}
-
-	var data map[string]string
-	err = json.Unmarshal(bytes, &data)
-
-	for k, v := range data {
-		block, _ := pem.Decode([]byte(v))
-		cert, err := x509.ParseCertificate(block.Bytes)
-		if err != nil {
-			return err
-		}
-		r.certificates[k] = cert
-	}
-	t, err := time.Parse("Mon, 2 Jan 2006 15:04:05 MST", resp.Header["Expires"][0])
-	r.expirey = t.Unix()
-	return nil
 }
 
 type IITAttestorConfig struct {
@@ -112,6 +54,11 @@ func (p *IITAttestorPlugin) Attest(ctx context.Context, req *nodeattestor.Attest
 		return &nodeattestor.AttestResponse{Valid: false}, err
 	}
 
+	if req.AttestedBefore {
+		err := cgcp.AttestationStepError("validation the InstanceID", fmt.Errorf("the InstanceID has been used and cannot be registered again"))
+		return &nodeattestor.AttestResponse{Valid: false}, err
+	}
+
 	identityToken := &cgcp.IdentityToken{}
 	_, err := jwt.ParseWithClaims(string(req.GetAttestedData().Data), identityToken, p.tokenKeyRetriever.retrieveKey)
 	if err != nil {
@@ -121,11 +68,6 @@ func (p *IITAttestorPlugin) Attest(ctx context.Context, req *nodeattestor.Attest
 
 	if identityToken.Audience != audience {
 		err = cgcp.AttestationStepError("Audience claim in the token doesn't match the expected audience", err)
-		return &nodeattestor.AttestResponse{Valid: false}, err
-	}
-
-	if req.AttestedBefore {
-		err = cgcp.AttestationStepError("validation the InstanceID", fmt.Errorf("the InstanceID has been used and cannot be registered again"))
 		return &nodeattestor.AttestResponse{Valid: false}, err
 	}
 
@@ -140,6 +82,9 @@ func (p *IITAttestorPlugin) Attest(ctx context.Context, req *nodeattestor.Attest
 		err = cgcp.AttestationStepError("validation of the ProjectID", fmt.Errorf("the projectID doen't match the projectID whitelist"))
 		return &nodeattestor.AttestResponse{Valid: false}, err
 	}
+
+	p.mtx.Lock()
+	defer p.mtx.Unlock()
 
 	resp := &nodeattestor.AttestResponse{
 		Valid:        true,
